+++ TimberCommons/IrrigationSystemUI/GoodConsumingBuildingFragmentPatch.cs
@@ -2,6 +2,7 @@
 // Author: igor.zavoychinskiy@gmail.com
 // License: Public Domain
 
+using System;
 using HarmonyLib;
 using IgorZ.TimberDev.UI;
 using Timberborn.GoodConsumingBuildingSystem;
@@ -15,6 +16,8 @@ namespace IgorZ.TimberCommons.IrrigationSystemUI {
 /// Harmony patch to show supply in days and hours.
 [HarmonyPatch(typeof(GoodConsumingBuildingFragment), nameof(GoodConsumingBuildingFragment.UpdateProgressBar))]
 static class GoodConsumingBuildingFragmentPatch {
+  const string SupplyRemainingLocKey = "GoodConsuming.SupplyRemaining";
+  static string _localizedSupplyRemainingTmpl;
 
   // ReSharper disable once UnusedMember.Local
   static void Postfix(bool __runOriginal, ILoc ____loc, Label ____hoursLeft,
@@ -22,7 +25,13 @@ static class GoodConsumingBuildingFragmentPatch {
     if (!__runOriginal) {
       return;  // The other patches must follow the same style to properly support the skip logic!
     }
-    ____hoursLeft.text = CommonFormats.DaysHoursFormat(____loc, ____goodConsumingBuilding.HoursUntilNoSupply);
+    if (_localizedSupplyRemainingTmpl == null) {
+      var original = ____loc.T(SupplyRemainingLocKey, "###");
+      _localizedSupplyRemainingTmpl = original.Substring(0, original.IndexOf("###", StringComparison.Ordinal)) + "{0}";
+    }
+    ____hoursLeft.text = string.Format(
+        _localizedSupplyRemainingTmpl,
+        CommonFormats.DaysHoursFormat(____loc, ____goodConsumingBuilding.HoursUntilNoSupply));
   }
 }