+++ TimberCommons/package/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.6 (started on 2/7/2024):
 * [Fix #33] Water valve doesn't handle contamination correctly.
+* [Fix #38] "Supply will last for" text disappeared in UI.
 * [Feature] Make WaterValve pausable.
 * [Feature] Show the tiles, irrigated by the `WaterTower`, as *green* on the map instead of "barely green, maybe dying".
 * [Enhancement] Show duration values below `0.01` as zero.